Minor Bugfix
My use of time.sleep and raw_input get in the way of the other thread
posting information to the console. Somehow..

At any rate. I'm going to need to fix that. In the meantime, I made a
"check" command that allows the other thread to post what it needs to
post.